Replacement for KeyboardInt.pm (RT #23947)
I've written a replacement for KeyboardInt.pm that both
facilitates what it is you want to do (try the password in
$ssh->config->get('pass')) but doesn't get in they way of
compliance with RFC 4256, or impose any limitations on any other
users of this package that may not enjoy a change in existing
behavior.  What I haven't written is any documentation for
it (yet).  If there's a chance that my replacement could just
become the default keyboard-interactive authentication method
implementation for Net::SSH::Perl, then I'd happily write up some
pod too, but here's the basic feature list and how it differs
from KeyboardInt.pm:

1) isn't limited to interactive mode

Incidentally, despite the name "keyboard-interactive", a
completely legitimate userauth info request may include a request
with empty strings for the name, and instruction fields and 0
num-prompts, the correct response for which is handled entirely
without any user interaction.  Given that the SSH authentication
protocol allows for multi-method authentication sequences, it's
completely possible for such a request to be part of a such a
sequence, and it would be silly to not support it just because
the client hadn't been configured as "interactive."  I consider
the current package to be buggy in this regard.

2) includes debugging

Figuring out where and why authentication is failing can be
tricky.  I just used the existing debugging framework and added
some appropriate debug calls.

3) implements control character filtering

Client's shouldn't blindly trust the text servers give them to be
safe for display, so I rip out any control characters per RFC
4251 section 9.2.

4) includes the ability to register info request handlers

Want to introspect the userauth info request and send an
automated reply to it if the request matches some set of
heuristics?  This the feature that lets you do that.  If you
don't register any handlers, this module acts pretty much the
same as KeyboardInt.pm acts now (modulo the above 3 features
anyway).

As an example of how you might use it to go about installing a
one-shot automatic password handler that will only be triggered
if the name and instruction strings in the userauth info request
message are empty, num-prompts is 1, the prompt looks like the
word "password" and we have a stored password for the current
user...

use Net::SSH::Perl;
use Net::SSH::Perl::Auth::KeyboardInteractive qw(add_kbdint_handler);
my $ssh = new Net::SSH::Perl->new($host, 'protocol' => 2);
add_kbdint_handler($ssh,
    sub {
        my ($auth, $name, $instruction, $lang, $prompts) = @_;
        return defined($auth->{'ssh'}->config->get('pass')) &&
               $name eq '' && $instruction eq '' &&
               @$prompts == 1 &&
               $prompts->[0]->[0] =~ /\Apassword\W*\z/i;
    },
    sub {
        my ($auth, $name, $instruction, $lang, $prompt, $echo) = @_;
        $auth->del_handler;
        return $auth->{'ssh'}->config->get('pass');
    }
);
$ssh-login(...