enable PTY support in SSH2 (RT #64517)
Currently, the use_pty support for SSH2 is disabled. I tried enabling it
but found that programs that read from STDIN often hang attempting to
read after consuming all input. I've put some work into fixing that.

My investigation indicates that the issue is due to the way that PTYs
are handled on the server side. The child process on the server side
never sees an EOF condition. When there is no PTY, the SSH server closes
the STDIN file descriptor, which causes EOF to return. Since STDIN and
STDOUT share a file descriptor when a PTY is allocated, the file handle
cannot be closed and the read blocks.

I believe that appending an 0x04 or two is actually the right thing to
do, based on _Advanced UNIX Programming_, section 4.2 (page 204):

=================================
The user can end a line in one of two ways. Most often, a newline
character is the terminator. The return key is pressed, but the device
driver translates the return (octal 15) to a newline (octal 12).
Alternatively, the user can generate an end-offile (EOF) character by
pressing Ctrl-d. In this case the line is made available to read as is,
without a newline terminator. One special case is important: If the user
generates an EOF at the start of the line, read will return with a zero
count, since the line that the EOF terminated is empty. This looks like
an end-of-file, and that’s why Ctrl-d can be thought of as an
end-of-file "character"
=================================

My testing indicates this is a workable solution. In addition, although
it *should* be the default behavior, I have set the "terminal modes"
string in the pty_req channel request to set VEOF (the character to be
treated as EOF) to 0x04, and provided a method to configure this value.