Fix Net::SSH::Perl::Key->extract_public()
The docs incorrectly specified the function took 2 parameters, the type
and the base64 encoded string.  This might be due to confusion on lvalue
list/context, vs rvalure list **Array** context;

    my @array = qw(a b c);
    my ($x) = @array;      # $x = 'a'
    my ($x) = qw(a b c);   # $x = 'a'
    my $x   = @array;      # $x = 3
    my $x   = qw(a b c);   # $x = 'c'

The new behavior correctly handles:

    my $k = Net::SSH::Perl::Key->extract_public('DSA', $key_string);

and:

    my $k = Net::SSH::Perl::Key->extract_public($key_string);

The only difference is when 2 args are sent and the first arg differs
from the detected key type, a warning is thrown.

The reason for supporting the single argument in this way is due the
interface being documented incorrectly.  This single argument version is
the only version that actually worked in the past, despite being
undocumented.  This patch maintains the hidden backwards compatibility
for others who may have code depending on this module.

Patch the extract_public() and new() to add the ability to extract the
comment field too.  Patch DSA/RSA dump_public() to write the comment if
it exists.

Create a t/07-keys.t which reads t/authorized_keys and makes sure that
extract_public() is working correctly.  The data for the tests comes
from the comment fields of the keys.  These can be add by:

    ssh-keygen -t <type> -b <bits> -C "type=RSA|DSA,size=<bits>"

For keys that you know should fail, adding a "FAIL" to the comment will
pass the test.