Handle authorized_keys option strings.
* Fix a bug in the Net::SSH::Perl::Key::RSA.
* Handle parsing SSH authorized_keys file options string prefixing.

This patch looks explicitly for the placement of the key typ
designations in each line it's passed.  It then uses that offset to
split the line into the options string and the keystring.  Currently,
the options string is just silently discarded, but it could be added to
the key object to allow for output like this in the dump_public()
routine:

    from="*.example.com" ssh-rsa <key> <comment>

Updated the tests to handle this as well.