fix master merge mess
Sorry all, that I messed up the master branch
with merges I tested too late and pushed too
fast to github.

This commit here is in fact a simple reverse
patch of the differences that piled up
between the v2 and the broken master so that
the v2 can become master again.

Later I can retry to resolve the pull requests
(PR #11 and #12) that I actually wanted to have
in here.